Se genera estructura MVC
