--- src/main/java/com/devsenior/nmanja/Book.java
@@ -1,30 +1,0 @@
-package com.devsenior.nmanja;
-
-public class Book {
-
-    private int id;
-    private String title;
-    private String author;
-
-    public Book(int id, String title, String author) {
-        this.id = id;
-        this.title = title;
-        this.author = author;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    
-    
-
-}
