+++ src/main/java/com/devsenior/nmanja/controller/LibraryService.java
@@ -1,0 +1,7 @@
+package com.devsenior.nmanja.controller;
+
+public class LibraryService {
+
+    
+
+}
