+++ src/main/java/com/devsenior/nmanja/model/Book.java
@@ -1,0 +1,39 @@
+package com.devsenior.nmanja.model;
+
+public class Book {
+
+    private String id;
+    private String title;
+    private String author;
+    private boolean isBorrowed;
+
+    public Book(String id, String title, String author) {
+        this.id = id;
+        this.title = title;
+        this.author = author;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public boolean isBorrowed() {
+        return isBorrowed;
+    }
+
+    public void setBorrowed(boolean borrowed) {
+        isBorrowed = borrowed;
+    }
+
+    
+    
+
+}
